// src/features/dashboard/models/vendor.model.ts
export interface VendorModel {
  // Id
  id: number;
  value: number;

  // Name
  name: string;

  // Vendor Contract
  contract: string;

  // Vendor Contract validate date
  contract_start_date: string;
  contract_end_date: string;

  // Vendor Contract Language
  contract_language: string;
}